Enable QuickFix actions to be selectively disabled
